+++ src/components/molecules/ProductCard/ProductCard.js
@@ -1,36 +1,7 @@
 import React, { useState, useEffect } from 'react';
 import { Card, Button } from 'react-bootstrap';
 
-export default ({ product }) => {
-  // localStorage.removeItem('cart');
-  const currCart = localStorage.getItem('cart');
-  let shopping = false;
-  if (currCart) {
-    shopping = JSON.parse(currCart);
-    // setCart(shopping.cart);
-  }
-  const [cart, setCart] = shopping ? useState(shopping.cart) : useState({});
-
-  const cartProduct = {
-    name: product.name,
-    price: product.price,
-    qty: 1,
-    stock: product.quantity,
-  };
-
-  const addToCart = () => {
-    cart[product.name] = cartProduct;
-    // cart.push(cartProduct);
-    setCart(cart);
-    localStorage.setItem('cart', JSON.stringify({ cart }));
-    console.log(cart);
-  };
-
-  // useEffect(() => {
-  //   console.log(cart);
-  // }, []);
-  // console.log(cart);
-
+export default ({ product, addToCart }) => {
   return (
     <Card className="sectionC-cards shadow p-3 mb-5 bg-white rounded">
       {/* <a href="/product"> */}
@@ -42,7 +13,7 @@ export default ({ product }) => {
           <p className="sectionC-card-price">R$ {product.price.toFixed(2)}</p>
         </Card.Text>
       </Card.Body>
-      <Button className="sectionC-card-button" variant="dark" onClick={addToCart}>
+      <Button className="sectionC-card-button" variant="dark" onClick={() => addToCart(product)}>
         Adicionar ao carrinho
       </Button>
     </Card>
