+++ src/components/pages/Home/Home.js
@@ -10,7 +10,7 @@ import {
 } from '../../molecules';
 import './Home.css';
 
-export default function Home() {
+export default function Home({ addToCart }) {
   const db = firebase.firestore();
   const [products, setProducts] = useState({});
   const [loaded, setLoaded] = useState(false);
@@ -26,23 +26,13 @@ export default function Home() {
     });
   }, []);
 
-  const { ecobag, stickers, tshirt, notebook } = products;
-
-  // example of product:
-  // ecobag.image
-  // ecobag.description
-  // ecobag.name
-  // ecobag.quantity
-  // ecobag.status
-  // ecobag.price
-
   return (
     <div>
       {loaded ? (
         <>
           <SectionA />
-          <SectionB product={tshirt} />
-          <SectionC products={products} />
+          <SectionB product={products.tshirt} />
+          <SectionC products={products} addToCart={addToCart} />
           <SectionD />
         </>
       ) : (
